requirements.txt
acme==1.12.0
appdirs==1.4.4
APScheduler==3.11.0
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
automationhat==0.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
boto3==1.13.14
botocore==1.20.0
buttonshim==0.0.2
Cap1xxx==0.1.3
certbot==1.12.0
certbot-dns-dnsimple==1.3.0
certbot-nginx==1.10.1
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
distro==1.5.0
dns-lexicon==3.3.17
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
Flask-Login==0.6.3
flask-mail==0.10.0
fourletterphat==0.1.0
future==0.18.2
gpiozero==1.6.2
guizero==1.1.1
gunicorn==23.0.0
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.6.4
ISStreamer==2.0.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
josepy==1.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
microdotphat==0.2.1
mock==4.0.3
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
mysql-connector==2.2.9
nose==1.3.7
nudatus==0.0.5
numpy==1.19.5
oauthlib==3.1.0
packaging==25.0
pantilthat==0.0.7
parsedatetime==2.6
parso==0.8.1
pbr==5.5.0
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
prompt-toolkit==3.0.14
psutil==5.8.0
PTable==0.9.2
pycairo==1.16.2
pycodestyle==2.6.0
pycups==2.0.1
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
PyICU==2.5
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyRFC3339==1.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.1
python-dotenv==1.1.0
python-prctl==1.7
pytz==2021.1
PyYAML==5.3.1
pyzmq==20.0.0
QScintilla==2.11.6
qtconsole==5.0.2
QtPy==1.9.0
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
s3transfer==0.3.4
scrollphat==0.0.7
scrollphathd==1.2.1
semver==2.10.2
Send2Trash==1.6.0b1
sense-emu==1.2
sense-hat==2.6.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
SoftLayer==5.8.5
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
thonny==4.0.1
tldextract==2.2.1
toml==0.10.1
tornado==6.1
touchphat==0.0.1
traitlets==5.0.5
twython==3.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==5.3.1
uflash==1.2.4
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wonderwords==2.2.0
wrapt==1.12.1
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
